ref(metrics): Rename metrics extraction module
