+++ src/sentry/incidents/logic.py
@@ -49,7 +49,7 @@ from sentry.snuba.entity_subscription import (
     get_entity_key_from_query_builder,
     get_entity_subscription_from_snuba_query,
 )
-from sentry.snuba.metrics.metric_extraction import is_on_demand_query
+from sentry.snuba.metrics.extraction import is_on_demand_query
 from sentry.snuba.models import SnubaQuery
 from sentry.snuba.subscriptions import (
     bulk_create_snuba_subscriptions,
